Add patches-dev feature
This adds functionality for a patches-dev key in the extra array in a composer
root file. Specific dev patches can be installed when composer is in dev mode.
This is useful to patch your package locally when there's a conflict in a
development environment that should not be patched in production.

Usage:
{
    "extra": {
        "patches-dev": {
            "This is a dev patch": "./path/to/path"
        }
    }
}

When the --no-dev argument is passed to composer, patches in the patches-dev
array are ignored.